Project conversion to typescript
