--- assets/src/components/Header.jsx
@@ -1,17 +1,0 @@
-const Header = () => (
-  <header className="main-header" id="header">
-		<div className="bg-color">
-			<nav className="nav navbar-default navbar-fixed-top">
-				<div className="container">
-					<div className="col-md-12">
-						<div className="navbar-header">
-							<a href="#header" className="navbar-brand">Word_2_PDF</a>
-						</div>
-					</div>
-				</div>
-			</nav>
-		</div>
-	</header>
-)
-
-export default Header;
